mcp-server/pyproject.toml
[project]
name = "legal-intelligence-mcp-hub"
version = "1.0.0"
description = "Unified MCP server for legal research - GPT Researcher, Court Listener, and Gemini tools"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Legal Intelligence Team" }
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "boto3>=1.35.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "moto>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["*.py", "clients/*.py", "tools/*.py", "tests/*.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.scripts]
legal-mcp = "server:main"

